fix: remove check for allowed extensions in directory mode
This check is breaking rendering of local images when -D flag is being
used. While the objective here seems to be to avoid rendering unexpected
files, probably for security reasons, it is not really the objective of
this app to be secure, since its use case is rendering local files in
trusted environments.